+++ types/interfaces/TagifyOptions.d.ts
@@ -1,0 +1,8 @@
+/// <reference types="yaireo__tagify" />
+/**
+ * Tag input UI options.
+ */
+export default interface TagifyOptions extends Tagify.TagifySettings {
+    maxChars?: number;
+    readonly?: boolean;
+}